多项重要底层优化：日志系统、Runner错误处理
- 优化日志系统，使之更符合逻辑，更利于维护，更稳定，不再出现 日志目录有时为空却还要使用 的bug。
- 优化Runner错误处理，使之可以处理内部Catch和Finally中的异常，并同时记录两个异常
- EBPUtility中添加Quote函数，并为所有shell参数中的路径添加引号，降低因路径存在空格而出错的可能性
- 添加SVNRevert.sh，并将程序内的Revert命令变为调用Shell文件
- 不再显示和记录对不存在的文件的删除操作，并将删除日志移动至单独的文件（和revert一起）

+++ Assets/_BuildPipeline/EazyBuildPipeline/PlayerBuilder/Editor/Extension/RunnerExtension.cs
@@ -159,13 +159,27 @@ namespace EazyBuildPipeline.PlayerBuilder
 
         private void RevertAllCopiedFiles()
         {
-            //删除所有已拷贝的文件
-            foreach (string file in allCopiedFiles)
+            //单独的日志文件
+            string logPath = Path.Combine(CommonModule.CommonConfig.CurrentLogFolderPath, "RevertFilesLog.txt");
+            using (var logWriter = new StreamWriter(logPath, true))
             {
-                Module.DisplayProgressBar("Delete Copied Files", file, 0, true);
-                File.Delete(file);
+                //删除所有已拷贝并存在的文件
+                Module.DisplayProgressBar("Start Delete Copied Files", 0, true);
+                logWriter.WriteLine("Start Delete Files"); logWriter.Flush();
+                foreach (string file in allCopiedFiles)
+                {
+                    if (File.Exists(file))
+                    {
+                        Module.DisplayProgressBar("Delete Copied Files", file, 0);
+                        File.Delete(file);
+                        logWriter.WriteLine("Deleted: " + file); logWriter.Flush();
+                    }
+                }
+                logWriter.WriteLine("End Delete Files"); logWriter.Flush();
+                Module.DisplayProgressBar("End Delete Copied Files", 0, true);
             }
             //还原目录
+            Module.DisplayProgressBar("Start Revert Copied Files", 0, true);
             List<UserConfig.PlayerSettings.CopyItem> copyList;
             switch (BuildPlayerOptions.target)
             {
@@ -182,16 +196,25 @@ namespace EazyBuildPipeline.PlayerBuilder
             string errorMessage = "";
             foreach (var item in copyList)
             {
-                string directoryName = Path.GetFileName(item.TargetPath);
-                Process p = SVNUpdate.Runner.ExcuteCommand("svn", "--non-interactive revert -R " + item.TargetPath,
-                    (object sender, DataReceivedEventArgs e) => { Module.DisplayProgressBar("Revert Copied Files in " + directoryName, e.Data, count++ % 1000f / 1000f, true); },
-                    (object sender, DataReceivedEventArgs e) => { errorMessage += e.Data + '\n'; }, null);
+                string title = "Revert Copied Files in " + EBPUtility.Quote(Path.GetFileName(item.TargetPath));
+                Module.DisplayProgressBar(title, "Start Revert " + EBPUtility.Quote(item.TargetPath), 0, true);
+
+                Process p = SVNUpdate.Runner.ExcuteCommand("/bin/bash",
+                    EBPUtility.Quote(Path.Combine(Module.ModuleConfig.ModuleRootPath, "SVNRevert.sh")) + " " +
+                    EBPUtility.Quote(item.TargetPath) + " " +
+                    EBPUtility.Quote(logPath),
+                    (object sender, DataReceivedEventArgs e) => { Module.DisplayProgressBar(title, e.Data, count++ % 1000f / 1000f); },
+                    (object sender, DataReceivedEventArgs e) => { errorMessage += e.Data + '\n'; },
+                    (object sender, EventArgs e) => { });
+
                 p.WaitForExit();
                 if (p.ExitCode != 0)
                 {
                     throw new EBPException("还原目录(" + item.TargetPath + ")时发生错误：" + errorMessage);
                 }
+                Module.DisplayProgressBar(title, "Finish!", 1, true);
             }
+            Module.DisplayProgressBar("End Revert All Copied Files", 1, true);
         }
 
         #endregion
